+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Migrate from Giant Swarm deprecated operatorkit framework to kube-builder.
+- Add finalizer to the silence.
 
 ### Fixed